+++ internetarchive/item.py
@@ -881,7 +881,8 @@ class Item(BaseItem):
                     retries_sleep: int | None = None,
                     debug: bool = False,
                     validate_identifier: bool = False,
-                    request_kwargs: MutableMapping | None = None) -> Request | Response:
+                    request_kwargs: MutableMapping | None = None,
+                    set_scanner: bool = True) -> Request | Response:
         """Upload a single file to an item. The item will be created
         if it does not exist.
 
@@ -1044,7 +1045,8 @@ class Item(BaseItem):
                                 file_metadata=file_metadata,
                                 access_key=access_key,
                                 secret_key=secret_key,
-                                queue_derive=queue_derive)
+                                queue_derive=queue_derive,
+                                set_scanner=set_scanner)
             return request
 
         if debug:
@@ -1134,7 +1136,8 @@ class Item(BaseItem):
                retries_sleep: int | None = None,
                debug: bool = False,
                validate_identifier: bool = False,
-               request_kwargs: dict | None = None) -> list[Request | Response]:
+               request_kwargs: dict | None = None,
+               set_scanner: bool = True) -> list[Request | Response]:
         r"""Upload files to an item. The item will be created if it
         does not exist.
 
@@ -1240,7 +1243,8 @@ class Item(BaseItem):
                                             retries_sleep=retries_sleep,
                                             debug=debug,
                                             validate_identifier=validate_identifier,
-                                            request_kwargs=request_kwargs)
+                                            request_kwargs=request_kwargs,
+                                            set_scanner=set_scanner)
                     responses.append(resp)
             else:
                 file_index += 1
@@ -1274,7 +1278,8 @@ class Item(BaseItem):
                                         retries_sleep=retries_sleep,
                                         debug=debug,
                                         validate_identifier=validate_identifier,
-                                        request_kwargs=request_kwargs)
+                                        request_kwargs=request_kwargs,
+                                        set_scanner=set_scanner)
                 responses.append(resp)
         return responses